Move PlansWithDefaults to its own file
